\name{Rcpp11-package}
\alias{Rcpp11-package}
\alias{Rcpp}
\docType{package}
\title{R / C++11 interface}
\description{The \pkg{Rcpp11} package provides C++ classes that
  greatly facilitate interfacing C or C++ code. 
}
\section{Introduction}{
  \pkg{Rcpp11} provides C++ classes to facilitate manipulation of a large 
  number of R data structures : vectors, functions, environments, ... 
}
\keyword{programming}
\keyword{interface}
